# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SiteSort API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for the company
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{projectId}/permits:
    get:
      operationId: listPermits
      tags: [permits]
      summary: List permits for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of permits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permit"
    post:
      operationId: createPermit
      tags: [permits]
      summary: Create a new permit
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePermitRequest"
      responses:
        "201":
          description: Permit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permit"
  /permits/{permitId}:
    patch:
      operationId: updatePermit
      tags: [permits]
      summary: Update a permit
      parameters:
        - name: permitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermitRequest"
      responses:
        "200":
          description: Updated permit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permit"
  /projects/{projectId}/photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List photos for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: logPhoto
      tags: [photos]
      summary: Log a compliance photo
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogPhotoRequest"
      responses:
        "201":
          description: Photo logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Project:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, on_hold, complete]
        startDate:
          type: string
          format: date
        targetEndDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        memberCount:
          type: integer
        alertCount:
          type: integer
        progressPercent:
          type: integer
      required: [id, companyId, name, address, status, startDate, createdAt, memberCount, alertCount, progressPercent]
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            recentActivity:
              type: array
              items:
                $ref: "#/components/schemas/ActivityItem"
          required: [recentActivity]
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        startDate:
          type: string
          format: date
        targetEndDate:
          type: string
          format: date
      required: [name, address, startDate]
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, on_hold, complete]
        targetEndDate:
          type: string
          format: date
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        description:
          type: string
        userId:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, type, description, createdAt]
...
    Permit:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          enum: [hot_works, confined_space, excavation, electrical, general]
        description:
          type: string
        responsibleUserId:
          type: string
        responsibleUserName:
          type: string
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, expiring_today, expired]
        documentUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, type, description, responsibleUserId, responsibleUserName, startDate, expiryDate, status, createdAt]
    CreatePermitRequest:
      type: object
      properties:
        type:
          type: string
          enum: [hot_works, confined_space, excavation, electrical, general]
        description:
          type: string
        responsibleUserId:
          type: string
        startDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        documentUrl:
          type: string
      required: [type, description, responsibleUserId, startDate, expiryDate]
    UpdatePermitRequest:
      type: object
      properties:
        description:
          type: string
        responsibleUserId:
          type: string
        expiryDate:
          type: string
          format: date
        documentUrl:
          type: string
    Photo:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        uploadedBy:
          type: string
        uploaderName:
          type: string
        photoUrl:
          type: string
        category:
          type: string
          enum: [progress, compliance, snag, safety_concern]
        description:
          type: string
          nullable: true
        zone:
          type: string
          nullable: true
        referenceNumber:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        takenAt:
          type: string
          format: date-time
      required: [id, projectId, uploadedBy, uploaderName, photoUrl, category, referenceNumber, takenAt]
    LogPhotoRequest:
      type: object
      properties:
        photoUrl:
          type: string
        category:
          type: string
          enum: [progress, compliance, snag, safety_concern]
        description:
          type: string
        zone:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required: [photoUrl, category]
...